backend/python/requirements.txt
torch
torchvision
ultralytics
pillow

